Convert ID to uint64 for minting

+++ payment/wallet.go
@@ -41,7 +41,7 @@ func CreateWallet(s user.Session, doMint bool) (*Wallet, error) {
 	}
 
 	if doMint {
-		mintedCoins := coin.MintUser(s.GetCurrentUser().UserID)
+		mintedCoins := coin.MintUser(uint64(s.GetCurrentUser().UserID))
 		for i := range mintedCoins {
 			cs.add(mintedCoins[i])
 		}
